fix: add document evaluation html report
